--- FE/apps/native/app/(tabs)/calendar/index.tsx
@@ -1,9 +1,0 @@
-import { Text, View } from "react-native";
-
-export default function CalendarScreen() {
-  return (
-    <View>
-      <Text className="text-blue-500">Calendar</Text>
-    </View>
-  );
-}
