--- FE/apps/native/app/(tabs)/mypage/index.tsx
@@ -1,9 +1,0 @@
-import { Text, View } from "react-native";
-
-export default function MypageScreen() {
-  return (
-    <View>
-      <Text className="text-blue-500">Mypage</Text>
-    </View>
-  );
-}
