--- FE/apps/native/app/(tabs)/program/index.tsx
@@ -1,9 +1,0 @@
-import { Text, View } from "react-native";
-
-export default function ProgramScreen() {
-  return (
-    <View>
-      <Text className="text-blue-500">Program</Text>
-    </View>
-  );
-}
